INTDEV-1090 module system: rename installer→applier, credentials→remote-url
The installer no longer fetches — that's the resolver's job — so the
name oversold the responsibility. Renamed to "applier" to match what
the function actually does (copy staged trees into place, persist
declarations). The credentials.ts file held only buildRemoteUrl, so
the name was broader than the contents — renamed to remote-url.ts.
