+++ tools/data-handler/src/modules/index.ts
@@ -15,9 +15,9 @@
 export * from './types.js';
 export * from './location.js';
 export * from './version.js';
-export * from './credentials.js';
+export * from './remote-url.js';
 export * from './source.js';
 export * from './inventory.js';
 export * from './resolver.js';
 export * from './orphans.js';
-export * from './installer.js';
+export * from './applier.js';
